restructure the code
