+++ configure.ac
@@ -1,8 +1,9 @@
-AC_INIT([lcal], [0.02])
+AC_INIT([lcal],[0.02])
 AC_CONFIG_SRCDIR([src/lcal.cpp])
 AM_INIT_AUTOMAKE
 CFLAGS="$CFLAGS -Wall -Werr -O2 -Wno-deprecated"
 AC_PROG_CXX
-AM_CONFIG_HEADER(include/config.h)
-AC_OUTPUT(Makefile src/Makefile)
+AC_CONFIG_HEADERS([include/config.h])
+AC_CONFIG_FILES([Makefile src/Makefile])
+AC_OUTPUT
