+++ include/helpers.hpp
@@ -1,0 +1,16 @@
+/*
+ * $Id: parsetemr.hpp,v 1.5 2004/01/20 19:54:33 ulrich Exp $
+ * Copyright (C) Ulrich Schwickerath
+ *  This file is part of lcal, a command line calculator 
+ *  Author: Ulrich Schwickerath, (C) 12/2000 CERN
+ *          ulrich.schwickerath@web.de
+ *  
+ *  This code comes with absolutly no warranty!
+ *  it should not be used for production purposes whatsoever
+ *
+ */
+
+#ifndef INCLUDE_HELPERS_HPP_
+#define INCLUDE_HELPERS_HPP_
+int remaining(const char *s, int* pos);
+#endif
