+++ src/Makefile.am
@@ -1,3 +1,3 @@
 bin_PROGRAMS = lcal
-lcal_SOURCES = lcal.cpp parseterm.cpp
+lcal_SOURCES = lcal.cpp helpers.cpp term.cpp sum.cpp fac.cpp pot.cpp
 lcal_LDADD = -lm
