+++ src/pot.cpp
@@ -26,11 +26,8 @@
 #include <stdlib.h>
 
 #include <iostream>
-#include "./parseterm.hpp"
-
-int remaining(const char *s, int* pos) {
-  return (strlen(s)-*pos);
-}
+#include "parseterm.hpp"
+#include "helpers.hpp"
 
 // implementation of Pot
 Pot::Pot() {
@@ -453,253 +450,4 @@ int Pot::isconst() {
     break;
   }
   return(0);
-}
-
-/* implementation details fac */
-Fac::Fac() {
-  Factors = 0;
-}
-Fac::Fac(const char * s, int* pos) {
-  ParseFac(s, pos);
-}
-Fac::~Fac() {
-  delete(Factors);
-  Factors = 0;
-}
-Fac * Fac::ParseFac(const char * s, int* pos) {
-  Pot * potenz = new Pot(s, pos);
-  Factors = new ListOf<Pot> (potenz);
-  while (remaining(s, pos) > 0 && s[*pos] == '^') {
-    (*pos)++;
-    potenz = new Pot(s, pos);
-    Factors->add_elem(potenz, '^');
-  }
-  return(NULL);
-}
-
-double Fac::value() {
-  Pot * potenz = (Factors->first_node())->element();
-  double val = potenz->value();
-  Factors->move_next();
-  Node <Pot>* i = Factors->current_node();
-  while (i != 0) {
-    val = pow(val, i->element()->value());
-    Factors->move_next();
-    i = Factors->current_node();
-  }
-  Factors->reset();
-  return val;
-}
-
-void Fac::print() {
-  Pot * potenz = (Factors->first_node())->element();
-  potenz->print();
-  Factors->move_next();
-  Node <Pot>* i = Factors->current_node();
-  while (i != 0) {
-    std::cout << "^";
-    i->element()->print();
-    Factors->move_next();
-    i = Factors->current_node();
-  }
-  Factors->reset();
-}
-
-int Fac::ispot() {
-  return(Factors->is_first() & Factors->is_last());
-}
-
-int Fac::isconst() {
-  int res = 0;
-  Node<Pot> * i = Factors->first_node();
-  while (i != 0) {
-    res += (i->element())->isconst();
-    Factors->move_next();
-    i = Factors->current_node();
-  }
-  return(res);
-}
-
-/* implementation part for sum */
-Sum::Sum() {
-  Summanden = 0;
-}
-Sum::Sum(const char * s, int* pos) {
-  ParseSum(s, pos);
-}
-Sum::~Sum() {
-  delete(Summanden);
-  Summanden = 0;
-}
-Sum * Sum::ParseSum(const char * s, int* pos) {
-  Fac * faktor = new Fac(s, pos);
-  Summanden = new ListOf<Fac> (faktor);
-  char action;
-  while (remaining(s, pos) >0 && (s[*pos] == '*' || s[*pos] == '/')) {
-    action = s[*pos];
-    (*pos)++;
-    faktor = new Fac(s, pos);
-    Summanden->add_elem(faktor, action);
-  }
-  return(NULL);
-}
-
-void Sum::print() {
-  Fac * faktor = (Summanden->first_node())->element();
-  faktor->print();
-  Summanden->move_next();
-  Node<Fac>* i = Summanden->current_node();
-  while (i !=0) {
-    if (i->action == '*') {
-      std::cout << "*";
-    } else {
-      std::cout << "/";
-    }
-    i->element()->print();
-    Summanden->move_next();
-    i = Summanden->current_node();
-  }
-  Summanden->reset();
-}
-
-double Sum::value() {
-  Fac * faktor = (Summanden->first_node())->element();
-  double val = faktor->value();
-  Summanden->move_next();
-  Node<Fac>* i = Summanden->current_node();
-  while (i != 0) {
-    if (i->action == '*') {
-      val = val * i->element()->value();
-    } else {
-      val = val / i->element()->value();
-    }
-    Summanden->move_next();
-    i = Summanden->current_node();
-  }
-  Summanden->reset();
-  return val;
-}
-
-int Sum::isfak() {
-  return(Summanden->is_first() & Summanden->is_last());
-}
-
-int Sum::ispot() {
-  if (Summanden->is_first() & Summanden->is_last()) {
-    return(((Summanden->current_node())->element())->ispot());
-  } else {
-    return(0);
-  }
-}
-
-int Sum::isconst() {
-  int res = 0;
-  Node<Fac> * i = Summanden->first_node();
-  while (i != 0) {
-    res += (i->element())->isconst();
-    Summanden->move_next();
-    i = Summanden->current_node();
-  }
-  return(res);
-}
-
-// implementation part of Term
-Term::Term() {
-  Terme = 0;
-}
-
-Term::Term(const char * s, int* pos) {
-  ParseTerm(s, pos);
-}
-
-Term::~Term() {
-  delete(Terme);
-  Terme = 0;
-}
-
-Term * Term::ParseTerm(const char * s, int* pos) {
-  Sum * summand = new Sum(s, pos);
-  Terme = new ListOf<Sum> (summand);
-  char action;
-  while (remaining(s, pos) > 0 && (s[*pos] =='+' || s[*pos] == '-')) {
-    action = s[*pos];
-    (*pos)++;
-    summand = new Sum(s, pos);
-    Terme->add_elem(summand, action);
-  }
-  if (remaining(s, pos) > 0 &&
-      s[*pos] != '+' &&
-      s[*pos] != '-' &&
-      s[*pos] != ')') {
-    std::cout << "Syntax error" << std::endl;
-    exit(1);
-  }
-  return(NULL);
-}
-
-double Term::value() {
-  Sum * summand = (Terme->first_node())->element();
-  double val = summand->value();
-  Terme->move_next();
-  Node<Sum> * i = Terme->current_node();
-  while (i != 0) {
-    if (i->action == '+') {
-      val = val + i->element()->value();
-    } else {
-      val = val - i->element()->value();
-    }
-    Terme->move_next();
-    i = Terme->current_node();
-  }
-  Terme->reset();
-  return val;
-}
-
-void Term::print() {
-  Sum * summand = (Terme->first_node())->element();
-  summand->print();
-  Terme->move_next();
-  Node<Sum> * i = Terme->current_node();
-  while (i != 0) {
-    if (i->action == '+') {
-      std::cout << "+";
-    } else {
-      std::cout << "-";
-    }
-    i->element()->print();
-    Terme->move_next();
-    i = Terme->current_node();
-  }
-  Terme->reset();
-}
-
-int Term::issum() {
-  return(Terme->is_first() & Terme->is_last());
-}
-
-int Term::isfak() {
-  if (Terme->is_first() & Terme->is_last()) {
-    return(((Terme->current_node())->element())->isfak());
-  } else {
-    return(0);
-  }
-}
-
-int Term::ispot() {
-  if (Terme->is_first() & Terme->is_last()) {
-    return(((Terme->current_node())->element())->ispot());
-  } else {
-    return(0);
-  }
-}
-
-int Term::isconst() {
-  int res = 0;
-  Node<Sum> * i = Terme->first_node();
-  while (i != 0) {
-    res += (i->element())->isconst();
-    Terme->move_next();
-    i = Terme->current_node();
-  }
-  return(res);
 }
